requirements.txt
Django>=5.0
python-dotenv
psycopg2-binary
cx_Oracle
Faker
